Renaming format and parse so they will not conflict with existing "format" option
