--- src/ca11y/lib/format.js
@@ -1,3 +1,0 @@
-export default function format(state, format, delimiter, date) {
-  return format.map((format)=> state[format]).join(delimiter)
-}
