+++ src/ca11y/lib/formatter.js
@@ -1,0 +1,3 @@
+export default function formatter(state, format, delimiter, date) {
+  return format.map((format)=> state[format]).join(delimiter)
+}
