+++ src/ca11y/lib/parser.js
