+++ src/ca11y/lib/validate.js
@@ -1,9 +1,9 @@
 export default function validate(options) {
-  if (!!options.parse && typeof options.parse !== 'function') {
-    throw new TypeError(`Ca11y: "options.parse" must be a function.`)
+  if (!!options.parser && typeof options.parser !== 'function') {
+    throw new TypeError(`Ca11y: "options.parser" must be a function.`)
   }
 
-  if (!!options.format && typeof options.format !== 'function') {
-    throw new TypeError(`Ca11y: "options.format" must be a function.`)
+  if (!!options.formatter && typeof options.formatter !== 'function') {
+    throw new TypeError(`Ca11y: "options.formatter" must be a function.`)
   }
 }
